+++ src/main/java/net/quantumfusion/dashloader/model/components/DashModelOverrideListBakedOverride.java
@@ -14,10 +14,10 @@ public class DashModelOverrideListBakedOverride {
     @Nullable
     @Serialize(order = 1)
     @SerializeNullable
-    public final Long model; // temp
+    public final Integer model; // temp
 
     public DashModelOverrideListBakedOverride(@Deserialize("conditions") DashModelOverrideListInlinedCondition[] conditions,
-                                              @Deserialize("model") @Nullable Long model) {
+                                              @Deserialize("model") @Nullable Integer model) {
         this.conditions = conditions;
         this.model = model;
     }
