+++ src/main/java/net/quantumfusion/dashloader/model/components/DashWeightedModelEntry.java
@@ -10,12 +10,12 @@ import net.quantumfusion.dashloader.mixin.accessor.WeightedBakedModelEntryAccess
 
 public class DashWeightedModelEntry {
     @Serialize(order = 0)
-    public Long model;
+    public Integer model;
 
     @Serialize(order = 1)
     public int weight;
 
-    public DashWeightedModelEntry(@Deserialize("model") Long model,
+    public DashWeightedModelEntry(@Deserialize("model") Integer model,
                                   @Deserialize("weight") int weight) {
         this.model = model;
         this.weight = weight;
